+++ Changelog.md
@@ -22,8 +22,7 @@ Changes:
   surrounding whitespace.
 - Standards compliant line ending handling. \n\r is not collapsed as a single
   line end anymore. Characters #x85 and #2028 are now handled as line end (unless
-  preceded by \r)
-  
+  preceded by \r)  
 - Update kotlinx.io support to 0.9.0, atomicfu to 0.31.0, kotlinx.serialization
   to 1.10.0, kotlinx.benchmark to 0.4.16, kotlin to 2.3.10, junit to 5.14.3.
 - Always expand entities in attribute values (causing an exception if the entity
@@ -38,6 +37,8 @@ Changes:
   but rather handles the strings directly (in a StringInOutBuffer) leading
   to speed and memory usage improvements (no additional buffering is needed
   for strings already in memory.
+- Per the XML standard processing instructions are actually allowed in the XML
+  body. Allow them here.
 
 Fixes:
 - Make ElementSerializer work better in an existing document context by